added multiple resources, prepared for merge
